+++ packages/types/src/index.ts
@@ -1,5 +1,5 @@
 export type { LogLevel, LogEntry, InternalLogEntry } from './log';
-export type { Span, SpanStatus, SpanAttributes } from './span';
+export type { Span, SpanStatus, SpanAttributes, SpanEvent } from './span';
 export type { Breadcrumb, BreadcrumbType } from './breadcrumb';
 export type { Transport } from './transport';
 export type { Integration, Client } from './integration';